requirements.txt
aiohttp==1.0.5
async-timeout==1.2.1
chardet==3.0.3
discord.py==0.16.7
multidict==2.1.5
pkg-resources==0.0.0
pytz==2017.2
websockets==3.3
